Add support for DragonFlyBSD. Use pthread_setaffinity_np on BSDs as well.
In most cases it suffices to just add || defined(__DragonFly__) to the
FreeBSD preprocessor checks. DragonFly only needs special handling for
retrieving the CPU core count, and it doesn't need the FreeBSD quirk
for the semaphore paths.

In both BSDs the pthread_{get/set}affinity_np() calls are defined in
the #include <pthread_np.h> header.

In DragonFlyBSD the pthread_{get/set}affinity_np() functions were only
recently added, so we need to check for __Dragonfly_version >= 400707.
Since this check needs the sys/param.h include, take this opportunity
to move the bunch of *BSD specific includes up a bit.

The pthread_{get/set}affinity_np() functions are already available in
FreeBSD since 8.1 and 9.0 releases, so it shouldn't be necessary to add
a version check for that.