fixed from PVS-Studio:
V666 Consider inspecting third argument of the function 'strncmp'. It is possible that the value does not correspond with the length of a string which was passed with the second argument. pcm.cpp 1104
V773 Visibility scope of the 'before2' pointer was exited without releasing the memory. A memory leak is possible. pcm-pcie.cpp 950
V773 Visibility scope of the 'after2' pointer was exited without releasing the memory. A memory leak is possible. pcm-pcie.cpp 950
V773 Visibility scope of the 'oldSocketStates' pointer was exited without releasing the memory. A memory leak is possible. pcmservice.h 344
V773 Visibility scope of the 'oldCoreStates' pointer was exited without releasing the memory. A memory leak is possible. pcmservice.h 344
V773 The function was exited without releasing the 'tmp' pointer. A memory leak is possible. pcm-iio.cpp 134