PCM: fix proper destruction
Prior to this commit, upon attempt at PCM object destruction the whole
program would've been aborted. This is due to termination of threads
without properly joining them.
As a result, the PCM object could not be destroyed once it was created,
or libpcm.so could not be unloaded once it was loaded.

This commit takes care of joining PCM threads upon object destruction.

Signed-off-by: Andrey Gelman <andrey.gelman@gmail.com>