PCM: shorten destruction time
CounterWidthExtender::UpdateThread:
Switch from simple sleeping to a conditional variable with timeout.
Conditional variable with timeout provides the same functionality as
sleep() most of the time, but can be woken up when the thread is being
joined, thus eliminating wait time for CounterWidthExtender object
destruction.

Notice,
the problem of long destructioin time did not exist prior to fixing
PCM destruction.

Signed-off-by: Andrey Gelman <andrey.gelman@gmail.com>