Add hires timestamp to CSV
Currently the PCM CSV has a timestamp with a 1 second resolution.
In order to correlate the PCM data with other data sources it
would be helpful to have a higher resolution timestamp.
This patch adds a timestamp which can be correlated with linux
perf or trace-cmd data or with ETW trace data.
perf and trace-cmd can be told to use the clock_gettime
CLOCK_MONOTONIC clock so that timestamp is used here for Linux.
Windows ETW uses GetSystemTimePreciseAsFileTime or
GetSystemTimeAsFileTime (on Win7 and before) so the patch uses
that timestamp on Windows. GetSystemTime is relatively slow so
after the 1st call to GetSystemTime* the QueryPerformanceCounter
API is used to get a time delta. The delta is added to the initial
GetSystemTime* value to get an absolute timestamp.