add spec file
Part of the effort to create a build flow is that we need a way to
generated the RPMs.
Today the only way to generate the RPMs is through rootfs_product_centos
which compiles all the code and place it under workspace_top/build/...
and then generate spec files on the fly from through the build process
and build the rpms.
Then there is python code that run on the CI that does these steps as
well and code in the CI that take these RPMs and place them in a folder
under lbrepo.

This change is mandatory to decouple the generation of the RPMs from the
create.sh scripts and the scripts that run in the CI as part of a
release test.

The idea is to have a single Makefile that will invoke all the `make rpm`
targets of the the repositories we want and will generate all the RPMs
that are part of the product.

These spec files are similar to the onse that where created by the
lb-rpm scripts but run without the need of create.sh and with out the
need of osmosis staging area - workspace_top/build/...

We add a single target called rpm that will generate the RPMs and will
place them under the dist folder.

issue: LBM1-8394