Don't use _exit in normal program termination
It is unnecessary to call _exit from the exit handler function.
Always using _exit breaks tools like gprof that require a normal exit call.
The sigINT_handler functions still call _exit, but this is just to preserve
the existing design for signal handling.
It may be desirable to call _exit for signals like SIGQUIT but not SIGINT.