c_example: fix rdtscp returning out of bounds CPU
This commit fixes an issue where Linux encodes the NUMA ID of a
given CPU in bits above 12 of IA32_TSC_AUX[31:0].
The code here was mistakenly including bit 12 and above of the
returned value in the CPU integer value.

When running the c_example on e.g. NUMA 1 of dual-NUMA platform, bit 12
was set causing an OOB value to be passed to e.g. pcm_c_get_cycles(),
resulting in a segfault.

Fix issue by removing the NUMA bits from the CPU calculation.

Signed-off-by: Harry van Haaren <harry.van.haaren@intel.com>