Make `WinPMem::LogError()` take a const pointer
This resolves a `const`ness issue where `winpmem.h` would pass string literals to `LogError()`, when that function expected a mutable pointer.

As this function only uses it as a `const` pointer, this just makes it take a `const` pointer instead.

This `const` error becomes visible in other projects that use PCM when they're compiling under MSVC with the `/permissive-` flag.