macOS: buildTopology in kernel now fills output buffer directly
Previously, buildTopology would allocate a buffer, fill it, and copy
it to the output buffer.  The buffer passed from userland is of the
same type that the kernel needs, and it is easier to fill it directly.
Also added comment clarifying need to keep types in user/kernel land
in sync.