scripts: topotop to show largest CPU share of each logical CPU
topotop --largest-share (or -l for short) shows new "CPU share" dot
next to CPU usage bar when CPU is under at least 25 % load. If this
dot is at the top, the process/thread that used most CPU time on that
CPU got at least 99 % of all the consumed CPU time.

This feature helps detecting by a glance if there too many threads
running on a CPU or a set of CPUs.

Signed-off-by: Antti Kervinen <antti.kervinen@intel.com>