pcm-bw-histogram.sh: expect `pcm-memory` to reside in same folder, rather than ./
As pcm-bw-histogram.sh is installed (renamed to pcm-bw-histogram) alongside the pcm-memory binary in sbin/, `pcm-memory` should be found in the same folder as the script, rather than the current folder.
Without this patch, pcm-bw-histogram cannot be launched (as it cannot find ./pcm-memory), unless you cd to the sbin/ folder. This is inconvenient, as the script creates a temporary file, as well as an output .csv file, in the current folder.
With this patch, you can just launch pcm-bw-histogram from any folder, and it will find pcm-memory in the same folder where it's been installed.